Remove .utils in test, move code to existing util.py
- `mk_sample_dataset`
- `mk_sample_product`

were added to utils module, but there was already existing `util.py`, moving
these function into existing file instead of creating new module.
